Upload git url directly

+++ lib/chest/registry.rb
@@ -1,13 +1,14 @@
 require 'json'
 require 'rest_client'
+require 'uri'
 require 'fileutils'
-require 'chest/helpers/zipfile'
+require 'parseconfig'
+require 'pp'
 
 class Chest::Registry
   def initialize(token=nil, api: 'http://chest.pm/api')
     @token = token
-    # @api = api
-    @api = 'http://localhost:3000/api'
+    @api = api
   end
 
   def request_raw(method, path, params=nil)
@@ -22,70 +23,65 @@ class Chest::Registry
   end
 
   def request(method, path, params=nil)
-    case method
-    when :get
-      JSON.parse(request_raw(:get, @api + path, params).body)
-    when :post
-      JSON.parse(request_raw(:post, @api + path, params).body)
-    when :delete
-      JSON.parse(request_raw(:delete, @api + path, params).body)
-    end
+    params.merge! token: @token
+    response = request_raw(method, URI.join(@api, path), params)
+    return JSON.parse(response.body)
   end
 
   def fetch_package(package_name)
     request :get, "/packages/#{package_name}"
   end
 
-  def fetch_package_versions(package_name)
-    request :get, "/packages/#{package_name}/versions"
-  end
-
   def download_package(package_name, version='latest', path)
     Dir.mktmpdir do |tmpdir|
-      archive_path = File.join(tmpdir, 'package.zip')
-      unarchived_path = File.join(tmpdir, package_name)
-      Dir.mkdir unarchived_path
-      open(archive_path, 'wb') do |f|
-        f.write request_raw(:get, @api + "/packages/#{package_name}/versions/#{version}/download").body
-      end
-      Zip::File.open archive_path do |zip_file|
-        zip_file.each do |entry|
-          entry.extract(File.join(unarchived_path, entry.to_s))
-        end
-      end
-
-      if Dir.exist?(path)
-        FileUtils.rm_r path
+      manifest = fetch_package(package_name)
+      repo_url = manifest['repository']['url']
+      suc = system "git clone #{repo_url} #{tmpdir}"
+      if suc
+        FileUtils.cp_r tmpdir, path
       else
-        Dir.mkdir path
+        return false
       end
-
-      FileUtils.cp_r unarchived_path, path
     end
   end
 
   def publish_package(input_path=Dir.pwd)
-    chest_config = JSON.parse(open(File.join(input_path, 'manifest.json')).read)
+    # Parse manifest.json
+    unless Dir.glob('*.sketchplugin')[0]
+      return false
+    end
+    manifest_path = File.join(input_path, Dir.glob('*.sketchplugin')[0], 'Contents', 'Sketch', 'manifest.json')
+    unless File.exist?(manifest_path)
+      return false
+    end
+    manifest = JSON.parse(File.read(manifest_path))
 
+    # Parse README file
     readme_path = File.join(input_path, 'README.md')
     readme = File.exist?(readme_path) ? File.open(readme_path).read : ''
-    metadata = chest_config.merge readme: readme
-    ignore_config_path = File.join(input_path, '.gitignore')
-    ignored_files = File.exist?(ignore_config_path) ? File.open(ignore_config_path).read.split(/\r?\n/) : []
-    ignored_files |= ['.', '..', '.git']
+    manifest.merge! readme: readme
 
-    response = nil
-
-    Dir.mktmpdir do |tmpdir|
-      archive_path = File.join tmpdir, "#{chest_config['name']}.zip"
-      Zipfile.new(input_path, archive_path, ignored_files).write
-      response = request :post, "/packages", token: @token, metadata: metadata, archive: File.new(archive_path, 'rb')
+    # Parse repository
+    unless manifest['repository']
+      gitconfig_path = File.join(input_path, '.git', 'config')
+      repository = ParseConfig.new(gitconfig_path)['remote "origin"']
+      if repository
+        repository_uri = URI.parse(repository['url'])
+        case repository_uri.host
+        when 'github.com'
+          url = "https://github.com" + repository_uri.path
+        else
+          url = repository_uri.to_s
+        end
+        manifest.merge! repository: {type: 'git', url: url}
+      end
     end
 
+    request :post, "/packages", manifest: manifest
     return response
   end
 
   def unpublish_package(package_name)
-    request :delete, "/packages/#{package_name}", token: @token
+    request :delete, "/packages/#{package_name}"
   end
 end
